Extract manifest update function
